v1.0.4
Added $ variable passed into main anonymous function as an alias for jQuery so that $ works in the script if jQuery is run in compatibility mode.

Added code to resize and crop in browser using canvas if processInline is true and browser supports canvas. This will use canvas_resize.js (https://github.com/danschumann/limby-resize) if it's available to do the resizing as the default browser resize implementation is generally not very good.

generatedImageFormat option (default: 'jpeg') - string - file format of generated image - valid choices 'jpeg', 'png' or others that make valid image MIME types.

generatedImageQuality option (default: 80) - integer - jpeg compression level of generated image, ignored if image isn't a jpeg - valid from 0 - 100.

Added code to handle EXIF orientation when processing in browser - this will correctly rotate the image when it is selected to fix issues with it being the wrong orientation when cropped. NOTE: this needs exif.js to be included otherwise this functionality will be skipped.

Added ability to provide an image to overlay on top of the image being cropped:

imgOverlay option (default: false) - string - the HTML of the overlay that will be placed over the image to be cropped.

imgOverlayOpacity option (default: 1) - integer - the opacity of the image overlay, valid from 0 - 1.

Added undo crop control so the user can recrop the original image without having to reselect and potentially re-upload it.

Added undoControl option (default: true) - boolean - whether to show the undo crop control after cropping.

Modified the zooming code so that the user cannot upscale the image to stop pixellation unless the image is smaller than the crop area or allowUpscaleBeyondCropSize is set to true.

Changed scaleToFill to allowUpscaleBeyondCropSize option (default false) - boolean - croppic always scales image to fill the crop widget but if this is true will allow the user to zoom the image beyond original size.

Added processInlineFallback option (default: false) - boolean - if processInline is true and this is true and the browser doesn't support processInline then fallback to uploading image to the server rather than provide error message.

Added inForm option (default: false) - boolean - flag to tell croppic that the widget is already in a form so it doesn't wrap the widget in a form but instead finds the existing form.

Added fileInputName option (default 'img') - string - the name of the file input that croppic creates.

Added croppedImage option (default: '') - string - the URL of an initial cropped image to display when initialising the widget.

Added file extension validation when user selects a file.

Added validTypes option (default: ['jpg', 'png', 'gif']) - array of strings - an array of file extensions to be considered valid. The code will automatically add 'jpeg' to the test if it finds 'jpg'. Also the code will allow jpg, png or gif for processing in browser as the image uploaded to the site will always be generatedImageFormat.

Added onBeforeReset callback option - callback function called before croppic resets itself.

Added onReset callback option - callback function called after croppic resets itself.

Added addExtraUploadControls callback option - callback function called when croppic is creating the upload controls so controlling script can add any extra controls. The function is passed one argument which is the upload controls container div as a jQuery object.

Added loadPicture(image) function - loads an image to be cropped after the widget has been created. Pass in an string containing the image URL or data URI (we use this to load an image taken by the user with a webcam using webcam.js).

Added some "global" variables on the Croppic function for browser support test results so controlling script can use the results:

isCanvasSupported - browser supports canvas.getContext("2d").

isAjaxUploadSupported - browser supports Ajax upload.

isProcessInlineSupported - browser supports FileReader for process inline.

Added a hidden input to the fallback IFrame that is used when ajax upload isn't supported to fix an error in IE where it would show a prompt to download file when receiving the JSON response from the server. The hidden input has the name "fallbackIframe" and the value 1. The server code can then check for the existence of fallbackIframe in the POST variables and adds the "Content-Type: text/plain" header which makes IE handle the response correctly.

Need to rewrite the img_save_to_file.php and img_crop_to_file.php code to handle changes.